+++ odoo/local-src/bso_hr_validation/__manifest__.py
@@ -4,8 +4,12 @@
     'description': 'BSO specific Leaves & Expenses validation + date',
     'author': 'BSO',
     'website': 'https://www.bsonetwork.com',
-    'depends': ['base', 'hr_expense', 'hr_holidays'],
+    'depends': [
+        'hr_expense',
+        'hr_holidays',
+    ],
     'data': [
+        'security/res_groups.xml',
         'security/ir.model.access.csv',
         'views/hr_expense_sheet.xml',
         'views/hr_holidays.xml',